requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
convertdate==2.3.2
flatbuffers==1.12
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hijri-converter==2.1.1
holidays==0.11.1
idna==2.10
joblib==1.0.1
Keras-Preprocessing==1.1.2
korean-lunar-calendar==0.2.1
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.4
pip==21.1.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyMeeus==0.5.11
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
setuptools==49.2.1
six==1.15.0
sklearn==0.0
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wheel==0.36.2
wrapt==1.12.1
keras==2.4.3